requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==3.6.4.0
black==21.12b0
celery==5.2.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
Deprecated==1.2.14
Django==3.2.11
django-autoslug==1.9.8
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.12.4
drf-yasg==1.20.0
flake8==4.0.1
flower==1.0.0
humanize==4.9.0
idna==3.6
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.5
MarkupSafe==2.1.3
mccabe==0.6.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==9.0.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyMySQL==1.1.0
pytz==2021.3
redis==4.1.0
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
sqlparse==0.4.4
tomli==1.2.3
tornado==6.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
